refactor: change file name service to db
