+++ src/app/roadmap/create/page.tsx
@@ -1,4 +1,4 @@
-import { getCategories } from "@/features/category/server/service";
+import { getCategories } from "@/features/category/server/db";
 import RoadmapForm from "@/features/roadmap/components/RoadmapForm";
 
 export default async function CreateRoadmapPage() {
