+++ src/app/roadmap/edit/[externalId]/page.tsx
@@ -2,7 +2,7 @@ import { headers } from "next/headers";
 import { notFound, redirect } from "next/navigation";
 import { db } from "@/db";
 import RoadmapForm from "@/features/roadmap/components/RoadmapForm";
-import { getRoadmapWithSession } from "@/features/roadmap/server/service";
+import { getRoadmapWithSession } from "@/features/roadmap/server/db";
 import { Roadmap } from "@/features/roadmap/type";
 import { auth } from "@/lib/auth";
